requirements.txt
beautifulsoup4==4.12.2
discord.py==2.3.2
h5py==3.10.0
numpy==1.23.4
Pillow==10.2.0
python-dotenv==1.0.0
Requests==2.31.0
